requirements.txt
asgiref==3.8.1
Django==4.0
djangorestframework==3.15.1
pillow==11.0.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
sqlparse==0.5.1
tzdata==2024.2
